requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
matplotlib>=3.7.0
seaborn>=0.12.0
altair>=5.0.0
folium>=0.14.0
streamlit-folium>=0.15.0
langchain>=0.2.0
langchain-experimental>=0.0.60
langchain-groq>=0.1.0
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
tabulate>=0.9.0
supabase>=2.3.0
httpx>=0.28.1
httpcore>=1.0.9
xlrd>=2.0.1
pyxlsb>=1.0.10
python-pptx>=0.6.21
scikit-learn>=1.3.0
kaleido==0.2.1